Atualização de bugs e falhas/ path de correção FTP

+++ frontend/src/screens/Admin/ListaUsuarios/index.tsx
@@ -1,11 +1,12 @@
 import React, { useContext, useEffect, useState } from 'react';
-import { View, Text, FlatList, Image, StyleSheet, TouchableOpacity, Alert, TextInput } from 'react-native';
-import { api } from '../../../services/api';
+import { View, Text, FlatList, Image, TouchableOpacity, Alert, TextInput } from 'react-native';
 import Icon from 'react-native-vector-icons/Ionicons';
+import styles from './style';
 import { AuthContext } from '../../../context/AuthContext';
+import { api } from '../../../services/api';
 
 const ListUsers = () => {
-    const { signOut, user, } = useContext(AuthContext);
+    const { signOut, user } = useContext(AuthContext);
     const [users, setUsers] = useState([]);
     const [searchQuery, setSearchQuery] = useState('');
     const [filteredUsers, setFilteredUsers] = useState([]);
@@ -44,20 +45,20 @@ const ListUsers = () => {
             await api.delete('/me/excluir', {
                 data: { userId }
             });
-            alert('Usuário excluído com sucesso.');
+            Alert.alert('Usuário excluído com sucesso.');
             setUsers(users.filter(user => user.id !== userId));
             if (userId === user.id) {
                 signOut();
             }
         } catch (error) {
-            alert('Erro ao excluir usuário.');
+            Alert.alert('Erro ao excluir usuário.');
             console.error('Error deleting user:', userId, 'erro:', error);
         }
     };
 
     const updateUserRole = async (userId, isAdmin) => {
         try {
-            await api.put('/users/cargo', { userId: userId, isAdmin: !isAdmin });
+            await api.put('/users/cargo', { userId, isAdmin: !isAdmin });
             const updatedUsers = users.map(user => {
                 if (user.id === userId) {
                     return { ...user, isAdmin: !isAdmin };
@@ -75,10 +76,7 @@ const ListUsers = () => {
     };
 
     const handleUpdateCargoUser = (userId, isAdmin) => {
-        // Verifica se o usuário atual é admin
-
         if (userId === user.id) {
-            // Verifica se há outros administradores no banco de usuários
             const otherAdmins = users.filter(user => user.isAdmin && user.id !== userId);
             if (otherAdmins.length === 0) {
                 Alert.alert(
@@ -91,10 +89,7 @@ const ListUsers = () => {
                     "Confirmar atualização",
                     "Você tem certeza que quer revogar o seu cargo? Você será desconectado do aplicativo.",
                     [
-                        {
-                            text: "Cancelar",
-                            style: "cancel"
-                        },
+                        { text: "Cancelar", style: "cancel" },
                         { text: "SIM", onPress: () => updateUserRole(userId, isAdmin) }
                     ]
                 );
@@ -104,38 +99,38 @@ const ListUsers = () => {
                 "Confirmar atualização",
                 "Você tem certeza que quer atualizar o cargo deste usuário?",
                 [
-                    {
-                        text: "Cancelar",
-                        style: "cancel"
-                    },
+                    { text: "Cancelar", style: "cancel" },
                     { text: "SIM", onPress: () => updateUserRole(userId, isAdmin) }
                 ]
             );
         }
-    }
+    };
 
     const handleDeleteUser = (userId) => {
         if (userId === user.id) {
-            Alert.alert(
-                "Confirmar exclusão",
-                "Você realmente quer excluir a si mesmo? Você será desconectado do aplicativo.",
-                [
-                    {
-                        text: "cancel",
-                        style: "cancel"
-                    },
-                    { text: "SIM", onPress: () => deleteUser(userId) }
-                ]
-            );
+            const otherAdmins = users.filter(user => user.isAdmin && user.id !== userId);
+            if (otherAdmins.length === 0) {
+                Alert.alert(
+                    "Erro",
+                    "Não é possível excluir o único administrador do sistema. Promova outro usuário a administrador antes de se excluir."
+                );
+                return;
+            } else {
+                Alert.alert(
+                    "Confirmar exclusão",
+                    "Você realmente quer excluir a si mesmo? Você será desconectado do aplicativo.",
+                    [
+                        { text: "Cancelar", style: "cancel" },
+                        { text: "SIM", onPress: () => deleteUser(userId) }
+                    ]
+                );
+            }
         } else {
             Alert.alert(
                 "Confirmar exclusão",
                 "Você tem certeza que quer excluir este usuário?",
                 [
-                    {
-                        text: "Cancelar",
-                        style: "cancel"
-                    },
+                    { text: "Cancelar", style: "cancel" },
                     { text: "SIM", onPress: () => deleteUser(userId) }
                 ]
             );
@@ -159,6 +154,7 @@ const ListUsers = () => {
             );
         }
     };
+
     const refreshList = async () => {
         try {
             const response = await api.get('/users');
@@ -169,10 +165,8 @@ const ListUsers = () => {
         }
     };
 
-
     const renderItem = ({ item }) => (
         <View style={styles.userContainer}>
-
             <View style={styles.userInfoContainer}>
                 <Image source={{ uri: `${api.defaults.baseURL}/files/${item.profileImage}` }} style={styles.profileImage} />
                 <View style={styles.userInfo}>
@@ -215,116 +209,5 @@ const ListUsers = () => {
         </View>
     );
 };
-
-const styles = StyleSheet.create({
-    container: {
-        flex: 1,
-        padding: 20,
-        backgroundColor: '#f0f0f0',
-    },
-    searchContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        backgroundColor: '#fff',
-        borderRadius: 8,
-        paddingLeft: 8,
-        flex: 1,
-        height: 50,
-        marginLeft: 10,
-    },
-    searchIcon: {
-        marginRight: 8,
-    },
-    searchInput: {
-        flex: 1,
-        height: 40,
-        fontSize: 16,
-    },
-    userContainer: {
-        marginBottom: 20,
-        padding: 10,
-        backgroundColor: '#fff',
-        borderRadius: 8,
-        elevation: 2,
-    },
-    userInfoContainer: {
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    profileImage: {
-        width: 50,
-        height: 50,
-        borderRadius: 25,
-        marginRight: 15,
-    },
-    userInfo: {
-        flex: 1,
-    },
-    name: {
-        fontSize: 18,
-        fontWeight: 'bold',
-    },
-    email: {
-        fontSize: 16,
-        color: '#555',
-    },
-    deleteButton: {
-        backgroundColor: '#ff5252',
-        padding: 10,
-        borderRadius: 8,
-        alignSelf: 'flex-start',
-        marginBottom: 10,
-    },
-    addAdminButton: {
-        backgroundColor: '#007bff',
-        borderRadius: 8,
-        flexDirection: 'row',
-        alignItems: 'center',
-        paddingHorizontal: 40,
-        paddingVertical: 10,
-    },
-    removeAdminButton: {
-        backgroundColor: '#ff8c00',
-        paddingHorizontal: 40,
-        paddingVertical: 10,
-        borderRadius: 8,
-        flexDirection: 'row',
-        alignItems: 'center',
-    },
-    userRoleButtonText: {
-        color: '#fff',
-        fontWeight: 'bold',
-        marginLeft: 5,
-    },
-    roleButtonContainer: {
-        marginTop: 10,
-        flex: 1,
-        alignItems: 'center',
-    },
-    searchContainerRefresh: {
-        flexDirection: 'row',
-        alignItems: 'center',
-        justifyContent: 'space-between',
-        padding: 10,
-        backgroundColor: '#f2f2f2',
-        marginBottom: 20,
-    },
-    refreshList: {
-        padding: 10,
-        borderRadius: 25,
-        backgroundColor: '#e0e0e0',
-        shadowColor: '#000',
-        shadowOffset: { width: 0, height: 2 },
-        shadowOpacity: 0.25,
-        shadowRadius: 3.84,
-        elevation: 5,
-    },
-    refreshIcon: {
-        fontSize: 24,
-        color: '#606060',
-    },
-
-
-});
 
 export default ListUsers;
